Test constructing CSA directly - maximum test data lines: 4000, because of stack overflow caused by several too~ big arrays. Fixed this problem by using (malloc).

+++ HelperFunction.h
@@ -8,8 +8,8 @@ void _myStrLengthTest() {
     printf("section\\\\ myStringLength test\n");
     char str[] = {'1','2','3','\0'};
     char *ch = &str[0];
-    int strlength = 0;
-    for(int i = 0; i < 3; i++) {
+    long strlength = 0;
+    for(long i = 0; i < 3; i++) {
         printf("%c ", str[i]);
     }
 
@@ -25,9 +25,9 @@ void _myStrLengthTest() {
 /*
  Get the length of a string. (including the character '\0')
  */
-int myStrLength(char str[]) {
+long myStrLength(char str[]) {
     char *ch = &str[0];
-    int strlength = 0;
+    long strlength = 0;
 
     while(*ch!='\0') {
         strlength++;
